Some minor cleanup.
I've modified ImageProcessing.java to allow methods to manipulate the whole array using a for loop instead of parsing a file one at a time. This is so that it can integrate with other tools in the main method better imo. But there seem to be a lot of unnecessary code for a small program like this. I'm still trying to think of a way to make the code more efficient.