Fixed some methods
First, I've fixed my wrapperRemover so that I don't need to instantiate a new instance everytime I'd like to convert a file. (rpush down instance - removing constructor parameter) . Added chooseDirectory function and also realized I missed a line which is to write to byte array. Issue now is that it overwrites the files with the last file in the array.